+++ test/runtests.jl
@@ -2,7 +2,9 @@ using Test
 using OperationsResearchModels
 using OperationsResearchModels.Simplex
 
+
 include("testrandomkeyga.jl")
+include("testtravelingsalesman.jl")
 include("testjohnsons.jl")
 include("testutility.jl")
 include("testjump.jl")